feat: enhance createInstance to return AsyncDisposable instances and add tests for lifecycle management
